+++ docs/agents/task-logs/weekly/2026-02-22_01-25-01_perf-optimization-agent_completed.md
@@ -1,0 +1,27 @@
+# Weekly Perf Optimization Agent - Completed
+
+**Agent:** perf-optimization-agent
+**Date:** 2026-02-22
+**Cadence:** Weekly
+
+## Summary
+Executed a performance optimization on the Moderation Stage in the feed engine.
+
+### Tasks
+1.  **Diagnosis:** Identified that `moderationStage` was performing sequential moderation lookups for each item in the feed, causing repeated calls to `getActiveTrustedMutersForAuthor` and potentially expensive pruning checks.
+2.  **Benchmarking:** Created `benchmarks/moderation_batch_bench.js` to simulate feed processing.
+    - Baseline: ~454ms for 5000 items (500 authors).
+3.  **Optimization:**
+    - Refactored `js/services/moderationService.js` to extract static utilities to `js/services/moderationUtils.js` (reducing file size and complexity).
+    - Implemented `getModerationInfoForAuthors(authors)` in `ModerationService` to batch author lookups and prune aggregates in a single pass.
+    - Updated `js/feedEngine/stages.js` to use `getModerationInfoForAuthors` for batch retrieval.
+4.  **Verification:**
+    - Re-ran benchmarks: ~227ms average (~50% improvement).
+    - Verified logic with `npm run test:unit`.
+    - Lint checks passed (resolved file size growth by extraction).
+
+## Artifacts
+- `benchmarks/moderation_batch_bench.js`
+- `BASELINE.md`
+- `AFTER.md`
+- `js/services/moderationUtils.js` (New file)
